Several changes to enable being run on singletons

+++ wdl/IgvCramPlots.wdl
@@ -131,13 +131,17 @@ task runIGV_whole_genome_localize{
             head -n+1 ~{ped_file} > family_ped.txt
             grep -w ~{family} ~{ped_file} >> family_ped.txt
             python3.6 /src/renameCramsLocalize.py --ped family_ped.txt --scc ~{sample_crai_cram}
-            cut -f4 changed_sample_crai_cram.txt > crams.txt
             
-            while read sample crai cram new_cram new_crai
-            do
-                mv $cram $new_cram
-                mv $crai $new_crai
-            done<changed_sample_crai_cram.txt
+            # Check if there are valid samples in the renamed file
+            if [ -s changed_sample_crai_cram.txt ]; then
+                cut -f4 changed_sample_crai_cram.txt > crams.txt
+                
+                while read sample crai cram new_cram new_crai
+                do
+                    mv $cram $new_cram
+                    mv $crai $new_crai
+                done<changed_sample_crai_cram.txt
+            fi
 
             i=0
             while read -r line
@@ -204,16 +208,27 @@ task runIGV_whole_genome_parse{
             cat ~{varfile} | cut -f1-3 | awk '{if (($3-$2)+int(($3-$2)*1.5)>=~{igv_max_window}) print $1"\t"$2-~{buffer}"\t"$2+~{buffer} "\n" $1"\t"$3-~{buffer}"\t"$3+~{buffer};
                 else print $1"\t"($2-int(($3-$2)*0.25))-~{buffer}"\t"$3+int(($3-$2)*0.25)+~{buffer}}' | sort -k1,1 -k2,2n | bgzip -c > regions.bed.gz
             tabix -p bed regions.bed.gz
-            #localize cram files
-            while read sample crai cram new_cram new_crai
-            do
-                name=$( echo $new_cram|awk -F"/" '{print $NF}'|sed 's/.cram//g' )
-                gsutil cp $crai $( basename $cram | sed 's/\.cram$/.crai/g' )
-                export GCS_OAUTH_TOKEN=`gcloud auth application-default print-access-token`
-                samtools view -h -C -T ~{reference} -o $name.cram $cram -L regions.bed.gz -M
-                samtools index $name.cram
-            done<~{updated_sample_crai_cram}
-            ls *.cram > crams.txt
+            
+            # Initialize crams.txt to avoid error if no cram files are created
+            touch crams.txt
+            
+            # Check if there are samples in the family before processing
+            if [ -s ~{updated_sample_crai_cram} ]; then
+                #localize cram files
+                while read sample crai cram new_cram new_crai
+                do
+                    name=$( echo $new_cram|awk -F"/" '{print $NF}'|sed 's/.cram//g' )
+                    gsutil cp $crai $( basename $cram | sed 's/\.cram$/.crai/g' )
+                    export GCS_OAUTH_TOKEN=`gcloud auth application-default print-access-token`
+                    samtools view -h -C -T ~{reference} -o $name.cram $cram -L regions.bed.gz -M
+                    samtools index $name.cram
+                done<~{updated_sample_crai_cram}
+                
+                # Only create crams.txt if there are cram files
+                if ls *.cram 1> /dev/null 2>&1; then
+                    ls *.cram > crams.txt
+                fi
+            fi
 
             i=0
             while read -r line
@@ -242,7 +257,7 @@ task runIGV_whole_genome_parse{
         Array[File] pe_txt = glob("pe.*.txt")
         Array[File] pe_sh = glob("pe.*.sh")
         Array[File] varfile = glob("new.varfile.*.bed")
-        Array[File] cramfiles = glob("*cram")
+        Array[File] cramfiles = glob("*cram") 
         }
     }
